[Feature] erase TC파일분리
